Move the test to "RemoveRedundantProjectSuiteBase.
